% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_commits.R
\name{count_commits}
\alias{count_commits}
\alias{get_git_commit_count}
\title{Get Git commit count}
\usage{
count_commits(from, to = Sys.Date(), author = NULL, path = NULL)

get_git_commit_count(from, to = Sys.Date(), author = NULL, path = NULL)
}
\arguments{
\item{from}{Date to count commits from.}

\item{to}{Optional ending date if you want to count commits in a date range.}

\item{author}{Optional author of the commits.}

\item{path}{Optional path of your Git repository.}
}
\description{
Get your Git commit count in a single date or range of dates.
}
\examples{
count_commits(from = "2019-06-01")
}
\keyword{git}
