% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_git_commit_history.R
\name{get_git_commit_history}
\alias{get_git_commit_history}
\title{Get Git commit history}
\usage{
get_git_commit_history(
  from,
  to = Sys.Date(),
  author = NULL,
  path = NULL,
  raw = FALSE
)
}
\arguments{
\item{from}{Starting date}

\item{to}{Ending date}

\item{author}{Optional author of the commits}

\item{path}{Optional path of your Git reporitory}

\item{raw}{Whether the output is a raw string vector (TRUE), or a data frame (FALSE)}
}
\description{
Get your Git commit history as a data frame or raw text between two specified dates
}
\examples{
get_git_commit_history(from = "2019-06-01")
}
\keyword{git}
